requirements.txt
Flask
qrcode[pil]
Pillow
gunicorn
azure-storage-blob